+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
 ### Added
-- (Breaking) __HOST- appending to increase Security to cookies on supported browsers. On by default. This will Append __HOST- to the fornt of cookie names.
+- (Breaking) __HOST- appending to increase Security to cookies on supported browsers. On by default. This will Append __HOST- to the front of cookie names.
   This will break some old Sessions as it changes the cookies names.
 
 ## 0.8.0 (23. October, 2023)